// client/src/app/model/MstData.ts
/**
 * Contain master data used for simulation
 * @author vupt
 */

export class MstData {
    mstPaper: any;
    mstPacking: any;
    mstPaste: any;
    mstStamping: any;
    mstDieCutting: any;
    mstSurfaceTreatment: any;
    mstWindow: any;
    mstColor: any;
    mstShippingCompany: any;
    mstLamination: any;
    mstShippingCost: any;
    mstPaperHead: any;
    mstPaperNormal: any;
    mstDecorative: any;
    mstShapes: any;
    mstCartonFlute:any;
    mstCartonShipping:any;
    mstColorFlexo: any;
}